tracker creation functionality using Core Data added
